+++ CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Added
+### Changed
 
-- `prefix` option for baricadr server url
+- Replace host/port by url in config
 
 ## [0.1.2] - 2021-03-17